third_party/blink/web_tests/platform/generic/media/W3C/video/events/event_timeupdate_manual-expected.txt

calling play() on a sufficiently long video should trigger timeupdate event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
